requirements.txt
aiogram==3.*
aiosqlite
httpx
python-dotenv
ruff
bs4